+++ CHANGELOG.md
@@ -3,8 +3,9 @@ xxx
 - Fixed `RequestQueue.addRequest()` to gracefully handle invalid URLs
 - Renamed `RequestOperationInfo` to `QueueOperationInfo`
 - Added `request` field to `QueueOperationInfo`
-- **WARNING**: Deprecated parameter `timeoutSecs` of `Apify.call()` was removed use `waitSecs` instead.
-  Parameter `timeout` is used for actor run timeout.
+- **WARNING**: Parameter `timeoutSecs` of `Apify.call()` is used for actor run timeout.
+    For time of waiting for run to finish use `waitSecs` parameter.
+- **WARNING**: Parameter `memory` of `Apify.call()` was renamed to `memoryMbytes`.
 - Added `Apify.callTask()` that enables to start actor task and fetch its output.
 
 0.8.18 / 2018-10-30